rename to memgraph
